feat(metric_alerts): Backfill SnubaQueryEventType rows for SnubaQuery without any.
We want to make sure that all `SnubaQuery` rows have at least one `SnubaQueryEventType` row. Before
event types existed, we always had an implicit event type based on the dataset of the `SnubaQuery`,
so we can just backfill using this if there are no existing `SnubaQueryEventType` rows present.

There are only 3k or so rules, so this should be fine to run as part of deploy.